requirements.txt
diffusers==0.19.3
emoji==2.7.0
numpy==1.24.1
Pillow==9.3.0
Pillow==10.1.0
requests_futures==1.0.1
torch==2.0.1+cu118
transformers==4.30.2
